perf: ⚡️ the hoverId will used only for inspector elements
The hoveredId will be changed whenever a mouse is over and out of the
component. This causes the state to re-render everytime when the
hoveredId changes. So a seperate state is maintained for the hovering of
elements. HoveredId in the redux state will be used only for the
elements in inspector.